[docs] Add policy for external dependencies
We wrote this document very early on in the project, but it got lost in
one of the repository shuffles. The original version is at
https://fuchsia.googlesource.com/manifest/+/refs/heads/master/policy.md

I've updated the document slightly to match our current practice. For
example, we no longer allow any unpinned dependencies and we have
different policies for prebuilt and source dependencies.

Change-Id: I738e67cdf3ddf9d7759054f111fb3597c82091b0