[clock][amlogic-clk] Add support for clocks on dosbus
Some clock registers for video decoding/encoding are on the DOS bus, not
on HIU. Add support to the clock driver for modifying registers in
multiple mmio regions.

Also a lot of video decoder clocks are always switched at the same time,
so add support to the driver for setting/unsetting a mask in a single
operation.

Change-Id: Ic61a1a55b66be65547a67c84add0f91177a7b506