[sysmgr] Start diagnostics before services and apps.
Ensures appmgr has a chance to construct appropriate diagnostics
buffers before any consumers start.

The diagnostics component URL is added as a config field to avoid
a non-hermetic PackageResolver in the sysmgr integration tests.

Further, we repackage sysmgr and observer.cmx with the integration
test to avoid passing service config as args to work around fxb/46567.

Change-Id: I5fe06a04ae33a818137b5e9888a6b303dbea7023