[sound_player] use audio renderers only once
Due to fxb/41074, sound_player cannot rely on a renderer to play two
sounds in a row with different formats. This CL changes sound_player
so it uses a given audio renderer to play a single sound only, after
which it discards the renderer.

This CL may be reverted once 41074 is fixed, or it can be left as-is.

BUG: 41074

TEST: new unit test verifies that renderers are not reused
Change-Id: I0cdd590995c4eef32f01c53424481157910d4cfd