[components] Library support for scoped dynamic instances
This change adds an API to the fuchsia-component library that allows a
component instance to be created in a collection. The instance's
lifetime is owned by a ScopedInstance object, and the instance is
destroyed (asynchronously) when this object goes out of scope.

This API will be used by the migrated font_provider tests, and should be
useful for other tests that wish to create components in-process.

Change-Id: Ied6618d8acba0b47053f47fd53b0cb77969fb24a