[zircon][kernel] Fix buffer overflow in vmar.
Fix possible overflows in vm_address_region. We used a lot of pattern
like base+size, however, even though we guarantee base+size-1 would not
overflow, base+size would. In order to prevent this, we change base+size
to base+(size-1), which is the last byte of this region and compare
that.

Tested: see fxr/363918
Bug: 45740

Change-Id: I51db3c9a66c570275d927adebaf71276e8ac5d04