[gn sdk] Add script for starting the Fuchsia emulator
This script starts up the Fuchsia emulator, and is a wrapper of the
the fx emu implementation.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/femu.sh or bin/femu-test.sh from
that project.

Note that fx emu also has its own set of separate in-tree tests and usage,
so the tests here are done to confirm the arguments are as expected.

Change-Id: I93797da8e08baeb35ee4fb2e3febe6ba726ec739