[gn sdk] Add script for starting Fuchsia DevTools
This script downloads Fuchsia DevTools from cipd, and then starts it up.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/fdevtools.sh or bin/fdevtools-test.sh
from that project.

Change-Id: Id2f6c90f559da82dec47e99b5d7010e7a3d0b1c3