[audio_core] Optimize filter table layout
The filters access the filter table in (1 << frac_bits) strides, which
for floats and 13 fractional bits, is an access every 32kb in the
underlying std::vector. Since these lookups are on the hot-path for the
filters, these accesses will have a high data cache miss rate.

Optimize this by providing a new CoefficientTable class that groups the
underlying floats such that values are (1 << frac_bits) apart are
physically adjacent.

Using a 10s trace, this yields a 607% speed increase of
Filter::ComputeSampleFromTable. Data collected with:

  $ fx traceutil record \
        -categories audio,audio.debug,kernel:meta,kernel:sched \
        --duration 10s \
        -buffer-size 64

Query data for Filter::ComputeSampleFromTable for SincSampler (note this
trace point was manually added for the purpose of this analysis).

  Before:
    avg      8,209 ns
    count  647,692
    max     49,458 ns
    min      3,333 ns
    std      3,917 ns

  After:
    avg      1,351 ns
    count  642,240
    max     11,750 ns
    min        917 ns
    std        446 ns

Bug: 45585
Test: fx run-test audio_core_unittests
      fx run-test audio_mixer_unittests
Change-Id: I8b98398e9d0de88e6c476212d8e605742298f190