[i18n] Use intl_services in terminal.gni
Provides an implementation of `fuchsia.intl.*` for all clients that do
not use modular framework.

Modular framework has its own implementation, which is also functional but
is mutually exclusive with this one.  And as modular is going away, we
expect that eventually we will use this implementation throughout.

Canonicalizes the build targets in //src/intl along the way.

Bug: 44741, 45216
Change-Id: I112f659d7fdfd8ec85e367b9b84ba28ed7a00fb0