[netstack2] Disable/Enable NIC in stack.Stack
When an interface is brought up or down, enable or disable it in
stack.Stack. This removes the need for the integrator to keep track of
dynamically learned routers, on-link prefixes and auto-generated
addresses as the NDP dispatcher will receive invalidation events when a
NIC is disabled.

This change also removes a route that considers all IPv4 addresses as
on-link.

Bug: 45956
Fixed: 43404

Test:
- Existing tests continue to pass.
- Manually made sure that bringing a link down removes all default IPv6
  routers, on-link prefixes and auto-generated IPv6 addresses.
- fx run-test netstack_gotests -t netstack_test -- \
     -test.run TestStackNICEnableDisable

Change-Id: Idb3d7e1f112a5576f733af02e222f06e6d5f33ec