[sktool] Revert post-approval changes made in fxr/363889.
After OWNERS approval was given on fxr/363889, some further changes were
made in sktool that both refactored the style to be less like the rest
of the package (and therefore made the package harder to read) and
introduced a bug where a completed iterator still continues to increment
its index, potentially eventually leading to an overflow.

I discussed this with Tamir out of band.

Change-Id: I41e26ef028faa734a51ba54b0997ae8eebd54ded