[zircon][kernel] Optimize ASLR code.
This CL optimize the logic to pick a random VM address. We first
randomly pick a location and if we could find it, we return early.
In most cases, we have lots of available spots so there are high
chances that we would hit the short path. Based on test on a board
and use a product, the average time of CreateSubVmar function is
reduced by 30% after the change.

Tested: Manual test on a product on a board.
Change-Id: I3f8ca05151977108fa39fb1601a189271371a14d