[i18n] Handles test flakiness in intl_services
This change should handle the test flakiness in intl_services_test.
IIUC, the issue was that we set a test value, then read it
asynchronously without receiving an OnChange signal, which acknowledges
an update of the applied value.

This could cause a subsequent read to read the old value of the profile,
and not the one we have set.

While I'm here, I added `FIDLError` and `Timeout` methods to the test
fixture to make error checks easier, and to put an upper bound to the
wait in the test case if something goes wrong in the event processing.

Bug: 46196
Change-Id: I46e2b9f37893dcf19ff756b2ea26d8d82b92dc30