[fidl][llcpp] Do not close the channel before calling on_unbound
Clients of fidl::AsyncBind have no way to send FIDL events on the
channel given to the binding other than holding a reference to the
channel, for instance keeping an unowned channel or keeping a handle.
Without a way for clients to synchronize themselves against the
channel not being closed, it is impossible for them to guard against
a crash (depending on policy) if they need to access the channel from
a thread different from the one(s) where the binding is running.
This change makes on_unbound guarantee that the binding end of the
channel is not closed by AsyncBind so it is always given to the callback
if provided by the client. With this change, clients can synchronize
their usage of the channel against the channel close that only happens
after the client has been notified.

Test: runtests -t fidl-async-test -r 3000
Change-Id: I8c96c3a6db06e8ac1e176f0b96db9db2e4bd812b