[rust][fidl] Return epitaphs in EventReceiver
This change propagates non-PEER_CLOSED epitaphs to EventReceivers
as Err(Error::ClientChannelClosed(epitaph)), after which the stream
returns None and terminates.

This change makes it possible to wait for a channel to close and
react to the reason (epitaph) why it closed. This is needed to
implement Component Runners, which communicate through events and
epitaphs (fxb/44545).

Bug: 44545

Change-Id: Iee925ea60d3dcb73b0b4e3061ac67342a7cca34f