[wlan][policy] Use stash to persist configs
Add a stash implementation to use for persistently storing network
configs. We do not remove the legacy implementation yet. If the
saved network is a WPA2 or Open network, we will save it in the legacy
storage as well as the new stash storage. And if network storage is
erased, the old storage will be erased as well.

Bug: 35910
Tests: Modified and added tests:
       fx run-test wlancfg-tests
       Manually tested that saving a network persists after reboot.
       Manually test that if we go from an older build to a build that
       uses this change, our saved network will still work.
Change-Id: I2c7f20ceb0bc106416b85a5ecbf9289256036e07