[system-tests] Cancel sub-tasks if test times out
This changes the system-tests to clean up resources (like spawned
processes or goroutines) if the context is canceled. Previously, this means
if the pave timed out, the test would properly exit with a timeout error,
but the pave-zedboot.sh/pave.sh script wouldn't get killed. This appears to
confuse the test executor into waiting for that process to exit as well.
This would cause the test to tie up any test resources until we hit
the test runner timeout.

Change-Id: Ifa0b0d712136d2770572eff812fff8aa206e254a