[amlogic_decoder] Make tests runnable on infra
Currently a gn arg is used to cause the integration tests to be built
into the real driver and run on driver startup. We don't have any way of
setting this on the bots, and we don't have any way of checking whether
the tests succeeded on the bots.

This change causes the test driver to always be built (and installed)
when the correct group is included in the build. A new test executable
unbinds the real driver and binds the test driver in its place - if any
tests fail then the test driver will fail to bind and the test
executable can detect that.

The test is currently disabled by default because
H264.DecodeMalformedHang is hanging.

Test: local testing

Change-Id: I4c6e5e4448452de5581d257b906d1686440e7d85