[camera] Fix flake in virtual_camera_test
This change fixes a flake in the virtual_camera_test suite due to not
explicitly destroying wait objects prior to destroying their associated
loop. It also fixes an issue where the fake_camera_test suite was using
the virtual_camera_test binary instead of the proper fake_camera_test
binary. This caused the same flake to be attributed to multiple tests.

FIXED: 46693
FIXED: 46427
Change-Id: I369aec2e006fa8b784bbf794dc21ce26f8b0e8d3