[sysmem] Run operations on new async::Loop
The driver currently uses the devhost's async loop, which is an
implementation detail that shouldn't be relied on. This patch switches
it to create a new async loop and run everything (after bind) on that
loop. This change also makes teardown cleaner, since tearing down the
loop causes the FidlServer to tear itself down automatically, preventing
it from outliving the device.

Change-Id: Ia10ff5541d256a1f5acf4ba6a731e018aa0b1a9a