[appmgr] Propagate and synthesize events from components managed by runners
The out/diagnostics directory propagation for runners includes a retry.
The reason of this is that flutter fills the out/ directory *after*
serving it. Therefore we need to watch that directory to notify.

Sadly the PseudoDir exposed in the SDK (and used by flutter) returns ZX_ERR_NOT_SUPPORTED on Watch.
This solution is implemented in fxr/366977 pending watch support.

This CL is a temporary solution (see bug) and adds max_retries for out/diagnostics being available

Bug: 46803
Tested: dart test in fxr/366315, fx bugreport and existing tests.

Change-Id: I66e4de49fbb1d4fad5c2665c42c5295282ed7245