[components] Client lib hosts EventStream
Have the client library host the EventStream. This will allow support
for both async and sync events.

From the component_manager's perspective, events will be fire and forget
UNLESS they have a Handler, and in which case, the component_manager
will wait on the handler's response. Currently all events have handlers,
but in the future, some events may be async and may not have handlers.

Change-Id: Idf11ac77cb6165c0f40fdb8342a8030d1aef9c90