[fidlc] Write tests to verify parsing of spans is correct
This sets up a framework so that we can easily add test cases for any
given type of SourceElement. Given a SourceElement subtype and a vector
of source strings that have expected spans of the given type marked
with guillemets ('«', '»'), the test will parse the source and verify
that each of the provided spans matches a node in the raw AST.

This CL adds basic test cases for {Const, Enum, Bits}Declaration spans,
and future CLs will add more test cases.

The test gives helpful errors for each failing test case. For example,
given the following test case
```
{
  ElementType::ConstDeclaration,
  {
    R"FIDL(
library example;

«const» uint32 C_SIMPLE = 11259375;
«const uint32 C_HEX_S   = 0xABCDEF»;
«const uint32 C_HEX_L   = 0XABCDEF»;
    )FIDL"
  }
}
```

The output will be:
```
    Test case for type ConstDeclaration...	Failed
        ERROR: unexpected occurrence of type ConstDeclaration: «const uint32 C_SIMPLE = 11259375»
        ERROR: expected (but didn't find) span of type ConstDeclaration: «const»
```

Test: fidlc host test (added span_tests)
Bug: 44553
Change-Id: I90af7e1500cc4834a939051881cb5af2ead77fd7