[lsan] Fixes leaks in unistd.cc
This started as an investigation of a leak in intl_wisdom_tests.
Digging deeper, the leak was in fact in unistd.cc: readdir and closedir
implementations had fd_to_io(...) calls which were not paired with
fdio_release, causing a reference miscount, which in turn caused the
underlying allocated `fdio_t` never to be freed.

This fix seems to remove many similar leak findings across our code
base.  Passes on CI and integration GQ.

Bug: 46735
Change-Id: I98eee78d05553acd28aaabe4c4af8c0693fe78bb