[display] Ensure VSYNC notification is not called with invalid channel
If client closes and the channel is closed, there is a race condition
where a vsync will try and write to the invalid channel and cause a
policy crash exception.

This CL ensure channel remains valid by owning the channel received
during fidl unbound.

The new unit-test fails if channel is destroyed during fidl unbind. It
passes with this new CL
Bug: 42892
Test: runtests -t display-core-unittests-test -r 1000

Change-Id: If62d6271d06856f24ad315882549b19c85e39786