[components][reland] Library support for scoped dynamic instances
This change adds an API to the fuchsia-component library that allows a
component instance to be created in a collection. The instance's
lifetime is owned by a ScopedInstance object, and the instance is
destroyed (asynchronously) when this object goes out of scope.

This API will be used by the migrated font_provider tests, and should be
useful for other tests that wish to create components in-process.

This change is a reland of 365574, which failed to integrate due to a
race with another change.

Change-Id: Ic544141ac39f9b37fbb589eccdbc725b8ab36952