[PowerManager] Introduce MockNode interface
Create a MockNode interface that can be used to emulate the behavior of
a Node. This will reduce dependencies between nodes for testing and
allow for more ergonomic control and better visibility into the behavior
of the node under test. This change also moves the thermal simulator
based thermal policy integration test into its own file.

Change-Id: I05312284f55b5d3f53d9c1f0c8a2890c6be8a372