[reland][factory_reset] Separate CLI tool from prod package, clean up build
This separates the "fdr" command line tool from the production
factory_reset package and cleans up the build rules to be more
consistent in their references to the production package, configuration
data, and CLI tool. The exposed GN build targets in
//src/recovery/factory_reset/BUILD.gn are now:

//src/recovery/factory_reset - production package and configuration
//src/recovery/factory_reset:tools - CLI tools (currently just "fdr")
//src/recovery/factory_reset:tests - tests for factory_reset

This separation allows deciding whether to include the CLI tools on a
per-product configuration basis.

The core product now includes the production target and configuration
in the base package set and includes the tools in the universe set
through the //bundles:tools target.  The SDK tools bundle also
explicitly includes the factory reset tools target.  References to
factory_reset related targets in the //garnet tree are not longer used
and are removed.

This is a reland of 10e045112c3c4839940a324325dfeb3f17aed8ba which was
reverted in d68046be556a153296b0e70e4bac7b85a8f7c57e due to an
out-of-tree reference to the //src/recovery/factory_reset:configs
target. This commit makes the //src/recovery/factory_reset target
include the configuration target, which is what everything wants, and
also includes a forwarding target with the previous contents. Once this
lands and rolls we can remove the reference to the fowarding target,
then remove the forwarding target.

Bug: 45780
Change-Id: Iacdfb8ad103f313cb25cd658af4c0bf892633dbc