[pkg] Separate command line tools from production services
This separate the build and aggregation rules for the command line
tools used to debug and control the software delivery subsystem from
the production services.  By default, both tools and services are
available in the base package set of the core product.  Separating
these allows product configurations that do not want or need the tools
to remove them.

Bug: 46969
Change-Id: I014b9385d516d2b0f7b389f6930829af9de4a1e0