[shell] Add some AST niceties
We add the HasErrors method for detecting whether a node has errors
anywhere in its branches, and nonterminals for Identifier and
Expression, which will make certain parse outputs easier to recognize.

We also add downcasting methods so we can inspect the types of different
nodes and use their unique getters.

Change-Id: I6404b4bcebe8d9ba04df7fb8774ed7203e440bc1