[debugger] Add TLS support
TLS resolution is somewhat complex, and tied very closely to linker
implementation. To better separate those concerns, libc has grown some
special symbols that contain DWARF code that knows how TLS is set up in
the linker and can resolve offsets into the TLS section for us.

To use these, the dwarf evaluator grows a "subroutine" function that
allows us to switch to executing a different instruction stream, then
continue the previous stream when we're done. This way every time we see
DW_OP_form_tls_addr, we assemble and run a small DWARF subprogram to
evaluate its argument.

At the moment the symbols containing this DWARF info are in process
memory like normal symbols. mcgrathr@ would like them to be in an
unallocated section in the future, meaning they wouldn't get mapped in
and we'd have to pull them from the binary in our debug info repo. For
now, however, they are allocated, and this saves us a lot of
(additional) stitching together of subsystems which were never meant to
speak.

Fixed: 5543
Change-Id: I9a112176413c49aa53518fdf82beec98bb8d6f2d