[audio_core] integrate thermal agent
This CL integrates the thermal agent into audio_core. This implies no
behavior change until audio_core is configured to respond to thermal
events.

Most of the code concerns distributing effect config changes through
the output pipelines. In addition, there is code in AudioCoreImpl to
host and start the thermal agent and to connect it to the
AudioDeviceManager.

TEST: unit tests provided for most changes

Change-Id: I5bb05695721bea29dc0f6361b2b6b5c5372d560e