[kernel][clocks] Add an "auto start" option to kernel clocks.
In response to a media-team feature request, add a new creation flag
to allow users to create kernel clocks in a pre-started state.  See
docs for details, but the simple summary is that clocks created with
this flag set will start life as a clone of clock monotonic, and then
may be subsquently maintained.  Also update the docs and add new tests
for auto-started clocks.

Test: New unit tests.

Change-Id: I11f88f083eba3d48b300d1a002f0850a284fcb43