[debugger] Add suggestion to error message.
The zxdb and fidlcat tools both need to launch the debug agent.
Sometimes people forget to run "fx serve" which is required to get this
component and get a seemingly unrelated error about being unable to
launch the debug agent.

Adds a sentence about running fx serve to these errors.

Change-Id: I1438a03dc690e2d3c87bbff0a0c61c81a5695aee