[components][reland] Library support for scoped dynamic instances
This change adds an API to the fuchsia-component library that allows a
component instance to be created in a collection. The instance's
lifetime is owned by a ScopedInstance object, and the instance is
destroyed (asynchronously) when this object goes out of scope.

This API will be used by the migrated font_provider tests, and should be
useful for other tests that wish to create components in process.

This change is a reland of 367458, itself a reland of 365574. 367458
failed to integrate due to a compilation error in realm.rs. It's still
being investigated why this error wasn't caught by CQ. In the meantime,
I've manually verified that fuchsia-component-tests builds and runs.

Change-Id: I6be48333ae635abba38e2f143393c69c19d8ad59