[perftest] Add "--random-order" option to run perf tests in a randomised order
This randomises the order of the performance test cases within the
process.

I am not making this the default yet because I need to check what
effect it has on the perfcompare trybot.  It is possible that this
does not explore enough test orderings, which could lead to a higher
false warning rate from perfcompare.

The current reason for adding this is to help rule out the influence
of test ordering when gathering results from the RoundTrip_* tests
when I'm doing local testing of IPC performance.

Bug: 45802
Test: new unit test added to Zircon "perf-test" executable
 * ran unit test locally with this:
     fx set bringup.x64 --with-base //garnet/packages/tests:zircon
     fx build
     fx qemu -k
     /boot/test/sys/perf-test
 * also checked manually that test ordering varies when running this:
     fx set terminal.x64 + build + boot
     zircon_benchmarks -p --random-order --runs=10 --filter=RoundTrip
Change-Id: I2262599c6d86fd42cc1cc70caeb7b8da1fd43c7f