[vx] Package loader doesn't give exec VMO
The process loader needs an executable VMO to support
fuchsia::process::Resolver::Resolve().  Today it relies on the package
loader providing an executable VMO for the package data, but we can
instead have it pull that VMO out of the package directory at
resource_path().  Then our package loader doesn't need to provide an
executable VMO at all, and so it doesn't need to attempt to mark VMOs as
executable ever.  This in turn allows us to remove ambient executability
from sysmgr and a couple sysmgr integration tests, which previously
required it.

`run_test_component` previously used fuchsia.process.Resolver to load
the meta/*.cmx file for component tests.  This change causes that call
to fail, as the contents of meta/ are never valid executables.  Instead,
`run_test_component` should use `fuchsia.sys.Loader` to load the cmx
file.

Bug: 37917
Change-Id: Ia4315dc3123b108ba6929db9e6fd974f5890c56b