[unification] Move //zircon/system/ulib/ftl-mtd/test and some other stuff to the GN build
This change moves the binaries under ulib/ftl-mtd/test and utest/fvm-host
along with their supporting libraries ulib/ftl-mtd and ulib/fvm-host. The
fvm host tool is also included in the transition.

All these elements need to be moved at the same time as they all depend on
the fvm-host library whose name sadly clashes with the export mechanism for
ZN libraries as "fvm-host" in the GN build could both be construed as
"the fvm-host library in the target toolchain" or "the fvm library in the
host toolchain".

This is a reland of a previous change, making sure that the fvm tool is
properly accessible this time around.

Generated with: //scripts/unification/move_zn_element.py

scripts/unification/verify_element_move.py --reference local/initial.json:

<none>

Bug: 36139
Change-Id: I98a72c40d0c4ae7f2a0dfa3fd4dbb2c9d0003e27