[identity] Minor cleanup on getopts nomenclature.
getopts uses the following terminology in its parsing of command line arguments:
* a "flag" is the special case of something that can be supplied on the
  command line that doesn't require (or sometimes doesn't require) any
  additional information. The presence of the flag is all that matters
  (this is why flags can never be required).
* an "option" is both the general case of a thing that can be supplied
  on the command line and the special case of such a thing that must be
  followed by additional information. The option exists only so that the
  information following it can be interpreted.
* an "argument" is the additional information that follows an option.

So in the case of a typical unix command `ls --size --hide goma* /tmp`
* "size" is a flag
* "hide" is an option
* "goma*" is the argument to the hide option

We haven't always been fully consistent in the names we've used in our
own code working with these concepts in getopts, which might cause a
little discomfort for some readers.

This CL attempts to use consistent terminology.

Change-Id: Ie9e1417cc80a313482fba2e489fd7dc57fa2b5de