[scripts] Fix features output of Cargo files
The Cargo.toml files generated by this script include an array of
"features", which consists of UTF-8 strings. When python prints the
representation of a UTF-8 string it prefixes it with "u" (e.g. u'abcd').

This change ensures that features are encoded as ASCII and therefore
appear correctly when printed as a representation into Cargo.toml files.
Not doing this results in unparsable Cargo.toml files.

Change-Id: I68e0b1f91d0e8727449c263729aa6520f3899a32