[bt][tools] Add write command to bt-bredr-profile
Add write command to bt-bredr-profile, which can be used to write data
to a socket that has been created by the connect-l2cap command or the
OnConnected event.

The write command does not guarantee that all data will be written. It
will perform 1 write and report the actual number of bytes that were
able to be written.

Bug: 40025
Test: PTS L2CAP/COS/CFD/BV-09-C
1) connect-l2cap <peer-id> 1 b 48
2) write 0 😀😀😀😀😀😀😀😀😀😀😀😀
3) disconnect-l2cap 0

Change-Id: Ifcaab70867288a354ad9f11a0134f688200ad598