[carnelian] Add a dynamic render API
Add a dynamic rendering API that can use either
spinel or mold.

This appears as a new view mode in Carnelian, but the
intent is for all the other view modes to go away in favor
of this rendering method over a short time.

Testing: Ran all the samples under both Scenic and FrameBuffer.

Change-Id: I258b44ea462b9011c7f5126303df2838f72579b9