[docs] Tracing in Fuchsia
Cleaning up the tracing docs and re-organizing content.

The main approach that I took was to try to reduce the
amount of information that was in a single place and organize
the documents in three key categories:

Concepts
Guides
Reference

By doing this, I think this makes documents that are easier to follow.

By design, the adding tracing to code guide doesn't cover every tracing macro, so as not to oversaturate the documents. Please let me know what you think (most of the information from the original documents is still in the documents, but re-organized and rewritten).

Short-term todos for me:
  Create a diagram of the tracing system
  Change images that are used in converting-visualizing-a-trace
  Do an editorial pass on:
    trace-format
    trace-specification-file

Long-term todos for me:
  Try to auto-generate most of the reference documents.
  Cleanup the "Advanced" topic section in the tracing guides
  to flow with the rest of the guides.

Change-Id: I9779eeba26576ea0c9079bc04b12f243c4cdd382