[sl4f] add facade for regulatory-region-service
Implement a facade for the RegulatoryRegionService, so that
we can call it from ACTS tests.

Design note
-----------
The RegulatoryRegionService implements two FIDL protocols. Although
we could have separate facades for each protocol, I think we're
better off with a single facade for the whole service.

The reason I prefer a single facade is that the split in the RRS
itself is to provide granular access control. Since the ACTS
tests don't need granular access control, I think the simplicity of
a single facade wins out over the benefits of directly mirroring
the FIDL protocol structure.

Bug: 46401
Test: manual (see below)
Change-Id: Id1f0f7e32e737ea0b21cc302643a1e67662057d8

Manual test
-----------
1. Deployed this code to a device along with fxr/367138
2. act.py -c ~/configs/fuchsia_testbed.config -tc InvokeSetRegionTest
3. Verified that device showed the following log message:
   *** QUICHE: received region #"JP"

Change-Id: Id1f0f7e32e737ea0b21cc302643a1e67662057d8