[device-finder] Fix unavailable addrs error.
It shouldn't be an error if no candidate addresses are available for a
connection. If attempting to connect to IPv6-only interfaces (as is
possible in some configurations), this removes the error that would
arise when attempting to connect with IPv4 enabled (see example
configuration in the attached bug).

This also updates all errors returned in mdns.go to use the `%w` format
to take advantage of error wrapping.

Fixed: 46654

Change-Id: Ic5a1bdf3c22016c7289b887c56ddee34646808a6