[session] Clarify coordinates across systems, part 1
Input and scenic are just two examples of systems that want to use
different corrdinate units within Fuchsia. This CL creates a struct that
allows systems to set and retrieve coordinates in pixels, pips, or
millimeters.

This CL adds Position, Size, and ScreenCoordinate structs to prepare for
migrating Workstation to the new system.

Bug: 44272
Change-Id: I96b08953fe86a11e33ce507f98b80f878d4f377c