[platform-bus] Add PBus::SetBootloaderInfo
This new function allows the sharing of bootloader information with the
platform bus. This is useful to make particular runtime decisions.

This is part of a series of CLs to enable the paver to decide how to
pave a device, based on the needs of a bootloader.

The next CL will introduce the FIDL API through which the paver can
query the bootloader information.

Bug: 46181
Change-Id: I04f54aab924a24a9b56781eea5427b5e11d70f42