[components][migration] Convert fonts.cmx and tests to components v2.
fonts.cmx and its tests are converted to v2. This required some code changes for
font_provider_integration_tests to create v2 components in a collection instead of
using the v1 Launcher API, and separate manifests to be written for
different command-line args.

This change does not actually move fonts into the v2 topology. This will
occur in a follow-up change.

Change-Id: I952bcff89b94e01da056bad382171620e1661a14