[http] Warn about URL with IPv6 address
In a URL, a numeric IPv6 address is supposed to have square brackets
around it, so warn if that's encountered, because oldhttp doesn't handle
that.

We could fix this by using HostNoBrackets() instead of host(), but
oldhttp is deprecated and will go away in favor of fuchsia.net.http, so
it's not clear that fixing is worthwhile - the assert will hopefully
save time for anyone else who hits this case, and encourage moving to
fuchsia.net.http instead.

Tested: CQ
Change-Id: I8d4420acd1c3bef178f0c98a40d7466e7ea8fcb5