[cm][cmc] Consolidate CM serde types
Today, type validation is done in a few places:

- CMC validation and translation to CM JSON
- CM JSON deserialization and translation to FIDL
- FIDL validation
- FIDL translation to cm_rust types

We've re-defined what `Name`, `Path`, and `Url` are a few times
in the code, and we also include hand written checks in
cm_fidl_validator.

This change aims to consolidate these common types and keep validation
logic in a central place.

Change-Id: I27e8e21bfd4bff48c311c5e6a261fc32a6770a2b