[gn sdk] Parse XML integration/prebuilts to find correct aemu prebuilt version
Previously we just used the latest CIPD binaries, which may not be working.
Now, we use the $FUCHSIA/integration/prebuilts file, which contains the
version of the latest working version. We use an XML parser to extract this
during the generation of the GN SDK. Also, we now store the aemu prebuilts
in a directory based on the name, so that if aemu.version changes, it will
automatically use the new version.

Fixes fxb/43901

Change-Id: Ia531de54f87bc68b23cec52f150afd004e65283d