[botanist] Send keep-alive pings
Sometimes botanist's syslog ssh session disconnects silently without the
client being aware of it. To detect this, we send pings every 5 seconds
with a 20-second timeout, and close the client (and then try to
reconnect) if the ping fails.

Also, if the log_listener command exits due to an ssh connection
failure, the syslogger will now continuously try to reconnect until it
succeeds (or the context gets cancelled). Previously it would only try
to reconnect once before exiting, which caused issues for the system OTA
test. A connection error would occur when the device began shutting down
to receive the new image, and the reconnection attempt would happen
during paving, at which point netstack was not yet available, so the
reconnection would fail and the syslogger would exit prematurely.

Example of a build with a non-truncated syslog (run using
fx make-integration-patch):
https://ci.chromium.org/b/8887261502046993392

Bug: 44825
Change-Id: Ief8afaa86bfc9b41749a9b40fb5f40694b7c750b