[scenic] Adds CreateSession2 constructor.
Previously was: [scenic] Replaces CreateSession with CreateSession2.

This change uses CreateSession2 to build a session so it can take a Focuser as a
parameter.

Previously, this changed substituted the constructor with one that had an
equivalent signature to avoid breaking clients; This happened during a roll, so
this change was rolled back.

This change keeps the old constructor and introduces another one to solve this
by doing a soft transition: submit this change, move clients to use the new
constructor, delete the old constructor.

Change-Id: I9f6d0c2530c23346b0805796080a733babdfe7e2