[rust] Rebase paths within compile_3p_crates.
Prior to this change, clients of rust_build_args would
be provided with a rustc path relative to the root output directory,
and a Cargo path that was rebased on top of the build directory.
This would result in build failures if clients spawned processes with
a working directory that was not the same number of levels below the
root as the output directory.

Fixed: 47101

Change-Id: I36dc87903a23072ba9c3266f966608ba9c091fd4