[bt][bt-gap] Move peer id from peer Inspect node to property
Previously, the peer id was included in the node name: "peer 00000000000001".

Move the peer id to a property within the peer node, and use a unique
number in the node name:
peer_0:
    peer_id: 00000000000000001

This will make it easier to create filters for privacy/security review.

Bug: 45905
Test: in bt-gap-tests
host_dispatcher::tests::on_device_changed_inspect_state

Change-Id: I8e4160d55e805176e64f34a9bcf2d7e22294e32c