[blobstats] Only evaluate host test in host toolchain
The host_env environment gets confused when evaluated in non-host
toolchains and assumes that the current_cpu refers to the host's
architecture.  This restricts evaluation of the target to happen only in
the host toolchain.

Concretely, this fixes Mac builds targeting arm64 targets as there is
no valid Mac arm64 environment.

Bug: 47333
Change-Id: Iaecc0b650ad5823acba8e0b636623848a802925f