[netstack3] Allow assigning link-local IPv6 addresses to interfaces
This change allows clients external to the netstack to add link-
local IPv6 addresses to interfaces.

In order to do this, we also consolidate what used to be two
separate storage buckets - one for non-link-local addresses and one
for the single link-local address - into one, and update relevant
code to adapt to this change.

Test: Added `test_get_ip_addr_subnet` and
      `test_add_ip_addr_subnet_link_local`; run with:
      `fx run-test netstack3_tests -t netstack3_core_lib_test`

Fixed: 46400

Change-Id: I29bc2db00c4c0ba70d2a1cdaa21563b9d29656b4