[unification] Greatly improve the handling of exported host libraries.
Instead of hosting the host version under a new name (foo-host), keep the original
name (foo) and let the template in charge of integrating the library within the
GN build decide on where to source the metadata for the library based on
the current toolchain.

This makes dependent build files way more reasonable as well as simplifies
migrations by not requiring searches under two different names.

Bug: 36548
Change-Id: I560a02d49af13537d3841f57ca11664e3b7566c4