[amber] Validate repository board before downloading packages.
During a system update, update_manager currently download packages and
_then_ validates that the board of the update matches the board it is
running on.

update_manager determines which board is is running on by looking at the
file "/config/build-info/board". Unfortunately, the download phase will
download the package "board-info", which will replace the file
"/config/build-info/board", leading to system_updater incorrectly
believing that the incorrect repository is actually a match for the
current board.

More precisely, in a scenario where an update repository is serving
updates for the wrong board:

  1. updater_manager determines that an update is available, connects to
     the repository, and fetches all the packages.

  2. As part of this process, it fetches the package "board-info", which
     includes an incorrect version of the file
     "/config/build-info/board". The current updater_manager process
     continues to see the old copy of the file, though new processes
     will see the new version.

  3. update_manager compares the current system's
     "/config/build-info/board" against the board of the repository.
     Because it doesn't match, update_manager fails.

  4. When update_manager runs again, it now uses the new, incorrect
     version of "/config/build-info/board".

  5. This time, update_manager will compare "board" against the
     repository's board, and see that they match.

  6. The incorrect image is written to disk.

We resolve this by verifying the board prior to downloading any new
packages.

Fixed: 46246
Change-Id: I21286973cac776da2ff0983e14e6eba1ccf05ccf