[components] Create lifecycle integration test
This test leverages the lifecycle channel example from
//examples/components/basic to create an integration test. The
test starts that component and then asks component manager to
destroy it. The test then waits to observe the stop event
related to the example component.

Even though the lifecycle channel is not currently passed, this
test succeeds because ELF runner is not yet passing the lifecycle
channel and therefore kills the component when asked. If this
test stays green while we add the lifecycle channel behavior to
ELF runner that is success.

Bug: 45895

Change-Id: I874a4a93bc97f3c6a34b805fb8066be25bae2990