[unification] Always build arm64 boot shims.
This will allow the board drivers to get migrated to the GN build, which
cannot be done at the moment as they data-depend on their respective shims.
The way shims are integrated in the GN build is very fragile at the moment
and will require some significant refactoring that's best served by proceeding
with the migrations.

Bug: 36139
Bug: 47406
Change-Id: I77529c6fe96c37bf3d24654ff8de846e2ba89b50